feat(zenfs): add the feature of garbage collection
Based on the `ZonedBlockDevice::AllocateZone()`, I added the feature to
execute the garbage collection in the ZenFS.

The garbage collection excutes like below.

1. Before does the garbage collection sequence, reset the all Zone
   which held invalid item only.
2. Select victim based on the policy. you can see the specific policy in
   the `ZonedBlockDevice::ZoneSelectVictim()`.
3. If the current device status satisfies to execute the garbage
   collection, each victims file can move to other Zone
4. At last, victim zone is reset.

[NOTE] You must think that the garbage collection runs recursively.

Additionally, I also added hot/cold sepearation when item assign to the Zone. This
was implemented to use the `Env::WriteLifeTimeHint`(aka. hotness). For instance, if
the some item's hotness is `Env::WLTH_EXTREME` and this item assigns to
the specific Zone(assume that Zone A), it only can be assigned item
which hotness is `Env::WLTH_EXTREME`.