fix(zenfs): consider the invalid file's lifetime
Previous version didn't care about the invalid file's lifetime when they
are in the Hot/Cold seperation. However, the Hot/Cold sepeartion scheme
must support the seperation whenever it is valid or invalid.

So, I fix it through add the ZoneMapEntry class which is used in the
file_map_. And you can exploit this when you want to knwo the exact
size of the valid or invalid in the zone.