fix(zenfs): repair the active zone issue
I found that the active zone issue via to run the fillseq in db_bench.
That issue was occured by the `active_zone_--` in my garbage collection
logic. So, I checked it and found it didn't necessary in the logic.
Moreover, I also saw that the active resource zone is limited in the ZNS
specifications. For this reason, I removed that line.

However, this still have the problem to execute fillseq in db_bench.
If you run the fillrandom in db_bench, you can execute up to 9,000,000
keys in the 8GB ZNS. However, if you run the fillseq in db_bench, you
can only execute up to 6,000,000 keys.

To sum up, we need to fix that the garbage collection logic a little to
handle the fillseq in db_bench.