fix(zenfs): total empty zones to empty spaces
In previous version is operated based on the total the number of the
empty zones. However, I found that it had an issue.

That is "it doesn't consider zone which is not fully written".

As you know, empty zone means that is empty. In other words, it doesn't
care about the "how many zone is not fully used". For example, when I
saw the zone status via monitoring program, I found that the number of
zones were almost 0%. However, the number of the empty spaces(sectors)
remain a lot. So, I thought that it didn't make sense of the garbage
collection logic.

As a result, I changed total empty zones based system to empty spaces
based system.